correct my copy error so existing scripts modified
